requirements.txt
babel==2.18.0
certifi==2026.4.22
charset-normalizer==3.4.7
et_xmlfile==2.0.0
idna==3.13
lxml==6.1.0
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
requests==2.33.1
six==1.17.0
tkcalendar==1.6.1
typing_extensions==4.15.0
urllib3==2.6.3
